Extract file content validation partial
